test(testing): оставшиеся persistence и infra tests вынесены из RSSReaderTests.swift
tests для AppDependencies, FeedRepository и FolderRepository выделены в отдельные suite files, а общий RSSReaderTests.swift убран из test target
