+++ README.md
@@ -334,6 +334,7 @@
 - [x] вынести `Settings`/`AppSettings` persistence/service/presentation/state/controller/theme tests из `RSSReaderTests.swift` в отдельные suite files, чтобы app settings flow был разложен по repository/service, screen presentation/state, controller orchestration и theme policy;
 - [x] вынести `SourceManagement` service/app-flow/state/controller tests из `RSSReaderTests.swift` в отдельные suite files, чтобы source onboarding и source editing flow были разложены по service-layer, app-level launch context, screen state и screen controller orchestration;
 - [x] вынести оставшиеся `Sidebar` selection/filtering tests, shell action/app-flow tests и `FeedNormalization`/`SourceIconCache` tests из `RSSReaderTests.swift` в отдельные suite files, чтобы монолитный test file сузился до небольшого persistence/infra хвоста;
+- [x] вынести оставшиеся `AppDependencies` и repository persistence tests из `RSSReaderTests.swift` в отдельные suite files и удалить сам `RSSReaderTests.swift`, чтобы test target полностью перешёл на feature-/layer-oriented структуру без общего монолитного suite file;
 - [ ] зафиксировать целевой sync scope: в CloudKit должны синхронизироваться `Feed`, `Folder`, `ArticleState` и `AppSettings`, а `Article` и `FeedFetchLog` должны оставаться только локальным storage слоем;
 - [ ] провести audit CloudKit compatibility для `AppSettings`, `Feed` и `Folder`: проверить `@Attribute(.unique)`, relationship semantics и текущие invariants, которые могут блокировать SwiftData sync;
 - [ ] провести audit CloudKit compatibility для `ArticleState`, `Article` и `FeedFetchLog`: проверить `#Unique`, nonoptional relationships, delete rules и закрепить границу между sync-backed сущностями и локальным article cache;
