+++ RSSReaderTests/Infrastructure/AppDependenciesTests.swift
@@ -1,0 +1,20 @@
+import Testing
+@testable import RSSReader
+
+@Suite("Infrastructure / AppDependencies")
+@MainActor
+struct AppDependenciesTests {
+    @Test
+    func appDependenciesExposeSeparateFolderRepositoryWhenSwiftDataIsAvailable() throws {
+        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
+
+        #expect(harness.dependencies.folderRepository != nil)
+    }
+
+    @Test
+    func appDependenciesExposeSourceManagementServiceWhenSwiftDataIsAvailable() throws {
+        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
+
+        #expect(harness.dependencies.sourceManagementService != nil)
+    }
+}
