+++ RSSReaderTests/Repositories/FeedRepositoryTests.swift
@@ -1,41 +1,10 @@
 import Foundation
-import SwiftData
 import Testing
 @testable import RSSReader
 
+@Suite("Repositories / Feed")
 @MainActor
-struct RSSReaderTests {
-    @Test
-    func appDependenciesExposeSeparateFolderRepositoryWhenSwiftDataIsAvailable() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-
-        #expect(harness.dependencies.folderRepository != nil)
-    }
-
-    @Test
-    func appDependenciesExposeSourceManagementServiceWhenSwiftDataIsAvailable() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-
-        #expect(harness.dependencies.sourceManagementService != nil)
-    }
-
-    @Test
-    func folderRepositoryPersistsInsertedFoldersAndReturnsSortedList() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let repository = try #require(harness.dependencies.folderRepository)
-
-        _ = try repository.insert(Folder(name: "Archive", sortOrder: 2))
-        _ = try repository.insert(Folder(name: "News", sortOrder: 0))
-        _ = try repository.insert(Folder(name: "Tech", sortOrder: 1))
-
-        let folders = try repository.fetchAllFolders()
-        let techFolder = try repository.fetchFolder(name: "Tech")
-
-        #expect(folders.map(\.name) == ["News", "Tech", "Archive"])
-        #expect(folders.map(\.sortOrder) == [0, 1, 2])
-        #expect(techFolder?.sortOrder == 1)
-    }
-
+struct FeedRepositoryTests {
     @Test
     func feedRepositoryUpdatesFolderAssignmentThroughExplicitPersistencePath() throws {
         let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
